backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.27
aiomysql==0.2.0
cryptography==42.0.5
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
sentence-transformers==2.5.1
redis>=5.0.1
elasticsearch[async]>=7.17.9
pymilvus>=2.4.1
httpx==0.27.0
aiokafka>=0.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
loguru>=0.7.2
PyYAML>=6.0.1
pytest-asyncio>=0.23.5
respx>=0.20.2
Pillow>=10.2.0
python-magic>=0.4.27; sys_platform != 'win32'
python-magic-bin>=0.4.14; sys_platform == 'win32'
